asn1crt: move int2uint and uint2int to asn1crt.h
The prototype was defined in asn1crt_encoding.h but implemented in
asn1crt.c. Keep the prototype in the same place as where the type it
declares is defined.

This also clears up the issue where int2uint and uint2int is not
properly linked when the generated code is compied with g++